v1.0

+++ src/main/java/com/example/sews/repo/OperateLogRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface OperateLogRepository extends JpaRepository<OperateLog, Long> {
-    // Custom queries can be added here if necessary
+public interface OperateLogRepository extends JpaRepository<OperateLog, Integer
+        > {
 }
